Initialize stampers in migration office
