+++ src/main/common/entities/CMakeLists.txt
@@ -5,4 +5,5 @@ file(GLOB SOURCES "*.h" "*.cpp")
 add_library(${PROJECT_NAME} ${SOURCES})
 
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC concurrency)
+target_link_libraries(${PROJECT_NAME} LINK_PUBLIC definitions)
 target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
