+++ src/main/migration_booth/MigrationBooth.h
@@ -5,13 +5,11 @@
 #include "Resident.h"
 #include "Foreigner.h"
 #include "SIGINTHandler.h"
-#include "SIGUSRHandler.h"
 #include "PersonsQueue.h"
 #include "Stampers.h"
 #include "Police.h"
 #include "StatisticsCommunicator.h"
 #include "Logger.h"
-#include "Semaphore.h"
 
 #include <string>
 
@@ -30,7 +28,7 @@ class MigrationBooth {
         void attend_foreigner(Foreigner* foreigner);
 
     public:
-        MigrationBooth(const int stampers_numer, const bool debug, const std::string log_file);
+        MigrationBooth(const bool debug, const std::string log_file);
         void open();
         ~MigrationBooth();
